refactor: 가이드 컴포넌트 네이밍 일관성 준수
