+++ client/main/src/pages/Guide/GuidePage.tsx
@@ -1,17 +1,17 @@
-import ContactContent from '../../components/Guide/Contact/ContactContent';
-import CreatorGuideContent from '../../components/Guide/Creator/CreatorGuideContent';
+import ContactForm from '../../components/Guide/Contact/ContactForm';
+import CreatorGuide from '../../components/Guide/Creator/CreatorGuide';
 import DonatorGuide from '../../components/Guide/Donator/DonatorGuide';
 import FeeGuide from '../../components/Guide/Fee/FeeGuide';
 import GuideHeader from '../../components/Guide/Header/GuideHeader';
 import SignUpGuide from '../../components/Guide/SignUp/SignUpGuide';
-import { StyledTemplate } from './GuidPage';
+import { StyledTemplate } from './GuidPage.styles';
 
 const contentMap: Record<string, () => JSX.Element> = {
   '/guide/sign-up': SignUpGuide,
   '/guide/donator': DonatorGuide,
-  '/guide/creator': CreatorGuideContent,
+  '/guide/creator': CreatorGuide,
   '/guide/fee': FeeGuide,
-  '/guide/contact': ContactContent,
+  '/guide/contact': ContactForm,
 };
 
 const GuidePage = () => {
